improve: push image api
- fix: base64 options to nowrapping
- separate token types which is returned from api (AuthToken)
  and passing to each request (Credential)
