+++ src/system.rs
@@ -107,3 +107,18 @@ impl AuthRequest {
         }
     }
 }
+
+/// Type of the response of `/auth` api
+#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
+#[allow(non_snake_case)]
+pub struct AuthToken {
+    Status: String,
+    IdentityToken: String,
+}
+
+impl AuthToken {
+    #[allow(dead_code)]
+    fn token(&self) -> String {
+        self.IdentityToken.clone()
+    }
+}
